Utility Functions

MeshGraphNets.der_minmax — Function

der_minmax(path)

Calculates the minimum and maximum derivative for each target feature in the given dataset.

Arguments

path: Path to the dataset files.

Returns

Minimum and maximum derivative in training, validation and test set.

der_minmax(path, is_training)

Calculates the minimum and maximum derivative for each target feature in the given part of the dataset.

Arguments

path: Path to the dataset files.

is_training: Determines for which dataset the calculation should be done. True for train and validation set, false for test set.

Returns

Minimum and maximum derivative in the specified part of the dataset.

MeshGraphNets.data_meanstd — Function

data_meanstd(path)

Calculates the mean and standard deviation for each feature in the given dataset.

Arguments

path: Path to the dataset files.

Returns

Mean and standard deviation in training, validation and test set.

data_meanstd(path, is_training)

Calculates the mean and standard deviation for each feature in the given part of the dataset.

Arguments

path: Path to the dataset files.

is_training: Determines for which dataset the calculation should be done. True for train and validation set, false for test set.

Returns

Mean and standard deviation in the specified part of the dataset.